centralize rb_form helpers to streamline rails2/3 view handling
