+++ app/controllers/rb_releases_controller.rb
@@ -1,4 +1,5 @@
 include RbCommonHelper
+include RbFormHelper
 include ProjectsHelper
 
 # Responsible for exposing release CRUD.
